Added most recent PHP and HTML
The latest and greatest version of Mumbo, revived back from the dead from the most recent backup. There may be other branches that are similar, but this is what is in live right now. Current commit is none of the static files (which is where the images from posts are stored).